# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "InvenTrack API",
        "description": "REST API untuk aplikasi manajemen inventori InvenTrack.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Server"
        }
    ],
    "paths": {
...
        "/api/purchase-orders": {
            "get": {
                "tags": [
                    "Purchase Orders"
                ],
                "summary": "Daftar purchase order",
                "operationId": "d963d5094ce66a18f23b0c88d31bf893",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "pending",
                                "received",
                                "cancelled"
                            ]
                        }
                    },
                    {
                        "name": "date_from",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "date_to",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List purchase order"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Purchase Orders"
                ],
                "summary": "Buat purchase order baru",
                "operationId": "702e17b215989244f25a33183c97ea13",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "supplier_id",
                                    "order_date",
                                    "status",
                                    "details"
                                ],
                                "properties": {
                                    "supplier_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "order_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-04-12"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "pending",
                                        "enum": [
                                            "pending",
                                            "received",
                                            "cancelled"
                                        ]
                                    },
                                    "notes": {
                                        "type": "string",
                                        "example": "Catatan PO"
                                    },
                                    "details": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "product_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "quantity": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "unit_price": {
                                                    "type": "number",
                                                    "example": 100000
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Purchase order berhasil dibuat"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/purchase-orders/{purchase_order}": {
            "get": {
                "tags": [
                    "Purchase Orders"
                ],
                "summary": "Detail purchase order",
                "operationId": "a78d20881546f28ddac9e28515b25e7f",
                "parameters": [
                    {
                        "name": "purchase_order",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail purchase order"
                    },
                    "404": {
                        "description": "Tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Purchase Orders"
                ],
                "summary": "Update purchase order",
                "operationId": "5069a98963c44780456c3d0d49513208",
                "parameters": [
                    {
                        "name": "purchase_order",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "supplier_id",
                                    "order_date",
                                    "status",
                                    "details"
                                ],
                                "properties": {
                                    "supplier_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "order_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-04-12"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "received",
                                        "enum": [
                                            "pending",
                                            "received",
                                            "cancelled"
                                        ]
                                    },
                                    "notes": {
                                        "type": "string",
                                        "example": "Catatan PO"
                                    },
                                    "details": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "product_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "quantity": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "unit_price": {
                                                    "type": "number",
                                                    "example": 100000
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Purchase order berhasil diupdate"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Purchase Orders"
                ],
                "summary": "Hapus purchase order",
                "operationId": "21278f91aee6ec56f9c6b9c009676e78",
                "parameters": [
                    {
                        "name": "purchase_order",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Purchase order berhasil dihapus"
                    },
                    "404": {
                        "description": "Tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}